requirements.txt
openai-whisper
sounddevice
numpy
requests
python-dotenv
